Make the cleanup code check if the temp directory exists before trying to delete it

+++ Tests/test_PopGen_FDist.py
@@ -46,9 +46,12 @@ class AppTest(unittest.TestCase):
         self._copyfile('out.cpl', 'out.cpl')
 
     def tearDown(self):
-        for file in os.listdir(self.dirname):
-            os.remove(self.dirname + os.sep + file)
-        os.rmdir(self.dirname)
+        #Not sure how exactly, but its possible the temp directory
+        #may not (still) exist.
+        if os.path.isdir(self.dirname) :
+            for file in os.listdir(self.dirname):
+                os.remove(self.dirname + os.sep + file)
+            os.rmdir(self.dirname)
 
     def test_datacal(self):
         """Test datacal execution.
